+++ src/java/crud/model/OrderDetail.java
@@ -10,19 +10,19 @@ package crud.model;
  * @author edgarrenderos
  */
 public class OrderDetail {
-    private long id;
-    private long idOrden;
-    private long quantity;
+    private int id;
+    private int idOrden;
+    private int quantity;
     private double total;
 
-    public long getID() { return id; }
-    public void setID(long value) { this.id = value; }
+    public int getID() { return id; }
+    public void setID(int value) { this.id = value; }
 
-    public long getIDOrden() { return idOrden; }
-    public void setIDOrden(long value) { this.idOrden = value; }
+    public int getIDOrden() { return idOrden; }
+    public void setIDOrden(int value) { this.idOrden = value; }
 
-    public long getQuantity() { return quantity; }
-    public void setQuantity(long value) { this.quantity = value; }
+    public int getQuantity() { return quantity; }
+    public void setQuantity(int value) { this.quantity = value; }
 
     public double getTotal() { return total; }
     public void setTotal(double value) { this.total = value; }
